+++ serialisation/src/main/java/org/ehrbase/serialisation/walker/Context.java
@@ -20,36 +20,68 @@
 package org.ehrbase.serialisation.walker;
 
 import com.nedap.archie.rm.RMObject;
-import org.ehrbase.webtemplate.model.WebTemplateNode;
-
 import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.HashMap;
 import java.util.Map;
+import org.apache.commons.lang3.tuple.ImmutablePair;
+import org.apache.commons.lang3.tuple.Pair;
+import org.ehrbase.serialisation.walker.defaultvalues.DefaultValues;
+import org.ehrbase.webtemplate.model.WebTemplateNode;
 
 public class Context<T> {
 
-    private final Deque<WebTemplateNode> nodeDeque = new ArrayDeque<>();
+  private final Deque<WebTemplateNode> nodeDeque = new ArrayDeque<>();
 
-    private final Deque<RMObject> rmObjectDeque = new ArrayDeque<>();
+  private final Deque<RMObject> rmObjectDeque = new ArrayDeque<>();
 
-    private final Deque<T> objectDeque = new ArrayDeque<>();
+  private final Deque<T> objectDeque = new ArrayDeque<>();
 
-    private final Map<WebTemplateNode, Integer> countMap = new HashMap<>();
+  private final Map<WebTemplateNode, Integer> countMap = new HashMap<>();
 
-    public Deque<WebTemplateNode> getNodeDeque() {
-        return nodeDeque;
+  private Map<Pair<String, String>, Deque<WebTemplateNode>> filteredNodeMap;
+  private DefaultValues defaultValues;
+
+  public Deque<WebTemplateNode> getNodeDeque() {
+    return nodeDeque;
+  }
+
+  public Deque<RMObject> getRmObjectDeque() {
+    return rmObjectDeque;
+  }
+
+  public Deque<T> getObjectDeque() {
+    return objectDeque;
+  }
+
+  public Map<WebTemplateNode, Integer> getCountMap() {
+    return countMap;
+  }
+
+  public Deque<WebTemplateNode> getSkippedNodes(WebTemplateNode childNode) {
+    Deque<WebTemplateNode> skippedNodes = null;
+    if (this.filteredNodeMap != null) {
+      skippedNodes =
+          this.filteredNodeMap.get(
+              new ImmutablePair<>(childNode.getAqlPath(), childNode.getRmType()));
     }
+    return skippedNodes;
+  }
 
-    public Deque<RMObject> getRmObjectDeque() {
-        return rmObjectDeque;
-    }
+  public void setFilteredNodeMap(
+      Map<Pair<String, String>, Deque<WebTemplateNode>> filteredNodeMap) {
+    this.filteredNodeMap = filteredNodeMap;
+  }
 
-    public Deque<T> getObjectDeque() {
-        return objectDeque;
-    }
+  public Map<Pair<String, String>, Deque<WebTemplateNode>> getFilteredNodeMap() {
+    return filteredNodeMap;
+  }
 
-    public Map<WebTemplateNode, Integer> getCountMap() {
-        return countMap;
-    }
+  public DefaultValues getDefaultValues() {
+    return defaultValues;
+  }
+
+  public void setDefaultValues(DefaultValues defaultValues) {
+    this.defaultValues = defaultValues;
+  }
 }